# backend/prediction-service/src/main/java/com/anghel/investmenthelper/prediction/controller/TrainController.java
package com.anghel.investmenthelper.prediction.controller;

import com.anghel.investmenthelper.prediction.model.dto.training.TrainingModelRequestDTO;
import com.anghel.investmenthelper.prediction.model.dto.training.TrainingModelResponseDTO;
import com.anghel.investmenthelper.prediction.service.model.PredictionModelMetadataService;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/v1")
public class TrainController {

    private final PredictionModelMetadataService predictionModelMetadataService;

    public TrainController(PredictionModelMetadataService predictionModelMetadataService) {
        this.predictionModelMetadataService = predictionModelMetadataService;
    }

    @PostMapping("/trainings/train")
    @PreAuthorize("hasRole('ADMIN')")
    public ResponseEntity<TrainingModelResponseDTO> trainModel(@Valid @RequestBody TrainingModelRequestDTO request) {
        return ResponseEntity.ok(predictionModelMetadataService.trainModel(request));
    }

    @PostMapping("/internal/trainings/train")
    public ResponseEntity<TrainingModelResponseDTO> trainModelInternal(@Valid @RequestBody TrainingModelRequestDTO request) {
        return ResponseEntity.ok(predictionModelMetadataService.trainModel(request));
    }
}