# investment-helper-openapi.yaml
openapi: 3.0.3
info:
  title: Investment Helper API
  version: 1.0.0
  description: Full API documentation for the Investment Helper microservices backend. Public endpoints are intended to be
    called through the API Gateway. Internal endpoints are used for service-to-service communication.
servers:
- url: http://localhost:8080
  description: Local API Gateway
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequestDTO:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          example: password123
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        dateOfBirth: &id001
          type: string
          format: date
          example: '2024-05-15'
      required:
      - email
      - password
      - firstName
      - lastName
      - dateOfBirth
    RegisterResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
        role:
          type: string
          example: ROLE_USER
        enabled:
          type: boolean
          example: true
    LoginRequestDTO:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          example: password123
      required:
      - email
      - password
    LoginResponseDTO:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token.
        signature:
          type: string
          description: JWT token field used by older Postman collection versions.
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          example: 1800
    RoleDTO:
      type: object
      properties:
        role:
          type: string
          example: ROLE_ADMIN
          enum:
          - ROLE_USER
          - ROLE_ADMIN
      required:
      - role
    AuthUserResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        role:
          type: string
          example: ROLE_ADMIN
        enabled:
          type: boolean
    UserInputDTO:
      type: object
      properties:
        authUserId:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
          example: john.doe@example.com
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        dateOfBirth: *id001
      required:
      - authUserId
      - email
      - firstName
      - lastName
      - dateOfBirth
    UserUpdateDTO:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john.updated@example.com
        firstName:
          type: string
          example: Johnny
        lastName:
          type: string
          example: Updated
        dateOfBirth: *id001
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        authUserId:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth: *id001
        active:
          type: boolean
          example: true
...
    TrainingModelRequestDTO:
      type: object
      properties:
        ticker:
          type: string
          example: AAPL
      required:
      - ticker
    TrainingModelResponseDTO:
      type: object
      properties:
        ticker:
          type: string
          example: AAPL
        modelVersion:
          type: integer
          example: 1
        accuracy: *id002
        recordsUsed:
          type: integer
          example: 252
        trainedAt: *id003
        status:
          type: string
          example: SUCCESS
...
    ErrorDTO:
      type: object
      properties:
        message:
          type: string
          example: Validation failed
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        path:
          type: string
          example: /api/v1/users/1
        timestamp: *id003
paths:
  /api/v1/auth/register:
    post:
      tags:
      - Auth
      summary: Register a new auth user and create the linked user profile.
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponseDTO'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDTO'
  /api/v1/auth/login:
    post:
      tags:
      - Auth
      summary: Login and receive a JWT access token.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDTO'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDTO'
  /api/v1/auth/users/{id}/role:
    patch:
      tags:
      - Auth
      summary: Update an auth user's role. Requires ADMIN.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponseDTO'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleDTO'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
  /api/v1/internal/auth-users/{authUserId}/disable:
    patch:
      tags:
      - Internal
      summary: Disable an auth user from another service.
      responses:
        '204':
          description: No content
      parameters:
      - name: authUserId
        in: path
        required: true
        schema:
          type: integer
          format: int64
  /api/v1/internal/users:
    post:
      tags:
      - Internal
      summary: Create user profile from auth-service.
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInputDTO'
  /api/v1/users/{authUserId}:
    get:
      tags:
      - Users
      summary: Get user profile by auth user id.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
      security:
      - bearerAuth: []
      parameters:
      - name: authUserId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    patch:
      tags:
      - Users
      summary: Update user profile by auth user id.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateDTO'
      security:
      - bearerAuth: []
      parameters:
      - name: authUserId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    delete:
      tags:
      - Users
      summary: Deactivate user profile by auth user id.
      responses:
        '204':
          description: No content
      security:
      - bearerAuth: []
      parameters:
      - name: authUserId
        in: path
        required: true
        schema:
          type: integer
          format: int64
  /api/v1/users:
    get:
      tags:
      - Users
      summary: Get all users. Requires ADMIN.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
      security:
      - bearerAuth: []
...
  /api/v1/trainings/train:
    post:
      tags:
      - Training
      summary: Train a prediction model for a ticker. Requires ADMIN.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingModelResponseDTO'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainingModelRequestDTO'
      security:
      - bearerAuth: []
  /api/v1/internal/trainings/train:
    post:
      tags:
        - Internal
      summary: Trigger model training from another microservice.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainingModelRequestDTO'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingModelResponseDTO'
...